Basic setup for initializing sharded machines
In the sharded case, we have to recover from multiple machines.
We initialize the machine with a list of shards to recover
from as well as a map indicating which machines each key
belongs to.

This also starts the merge process for the parent input records.

This commit may be easier to view with -w.